% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualization.r
\name{Addposition}
\alias{Addposition}
\title{Add position information to fitted model}
\usage{
Addposition(obj, posList)
}
\arguments{
\item{obj}{A iSCMEBResObj object.}

\item{posList}{A list of two columns matrices, each element contains the position information of i-th tissue section.}
}
\description{
The function \code{Addposition} is used to add position information to the fitted model.
}
\examples{
data(iSCMEBObj_toy)
library(Seurat)
posList <- iSCMEBObj_toy@posList
resList <- iSCMEBObj_toy@resList
resList <- Addposition(resList, posList)
}
\seealso{
\code{\link{fit.iscmeb}}, \code{\link{iSCMEBResObj-class}}
}
