% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualization.r
\name{CalculateTSNE}
\alias{CalculateTSNE}
\alias{CalculateTSNE.iSCMEBObj}
\alias{CalculateTSNE.iSCMEBResObj}
\alias{CalculateTSNE.list}
\title{Run t-SNE dimensionality reduction.}
\usage{
CalculateTSNE(obj, reduction = NULL, n_comp = 3, seed = 1, toList = TRUE)

\method{CalculateTSNE}{iSCMEBObj}(obj, reduction = NULL, n_comp = 3, seed = 1, toList = TRUE)

\method{CalculateTSNE}{iSCMEBResObj}(obj, reduction = NULL, n_comp = 3, seed = 1, toList = TRUE)

\method{CalculateTSNE}{list}(obj, reduction = NULL, n_comp = 3, seed = 1, toList = TRUE)
}
\arguments{
\item{obj}{A iSCMEBObj object or a iSCMEBResObj object or a list of low dimension embeddings.}

\item{reduction}{The name of embeddings to be used to calculate t-SNE. If reduction is null, the last added one is used for plotting.}

\item{n_comp}{An optional positive integer, specify the number of features to be extracted.}

\item{seed}{A random seed to be used.}

\item{toList}{An indicator of whether convert the UMAP representation to a list.}
}
\value{
Returns a revised iSCMEBObj object.

Returns a revised iSCMEBResObj object.

Returns the UMAP matrix.
}
\description{
The function \code{CalculateTSNE} is used to run t-SNE dimensionality reduction.
}
\examples{
data(iSCMEBObj_toy)
library(Seurat)
iSCMEBObj_toy <- CalculateTSNE(iSCMEBObj_toy, reduction="iSCMEB")
## resList <- CalculateTSNE(iSCMEBObj_toy@resList, reduction="iSCMEB")
}
\seealso{
\code{\link{iSCMEBObj-class}}, \code{\link{fit.iscmeb}}, \code{\link{iSCMEBResObj-class}}
}
