% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualization.r
\name{CalculateUMAP}
\alias{CalculateUMAP}
\alias{CalculateUMAP.iSCMEBObj}
\alias{CalculateUMAP.iSCMEBResObj}
\alias{CalculateUMAP.list}
\title{Run UMAP dimensionality reduction}
\usage{
CalculateUMAP(obj, reduction = NULL, n_comp = 3, seed = 1, toList = TRUE)

\method{CalculateUMAP}{iSCMEBObj}(obj, reduction = NULL, n_comp = 3, seed = 1, toList = TRUE)

\method{CalculateUMAP}{iSCMEBResObj}(obj, reduction = NULL, n_comp = 3, seed = 1, toList = TRUE)

\method{CalculateUMAP}{list}(obj, reduction = NULL, n_comp = 3, seed = 1, toList = TRUE)
}
\arguments{
\item{obj}{A iSCMEBObj object or a iSCMEBResObj object or a list of low dimension embeddings.}

\item{reduction}{The name of embeddings to be used to calculate UMAP. If reduction is null, the last added one is used for plotting.}

\item{n_comp}{An optional positive integer, specify the number of features to be extracted.}

\item{seed}{A random seed to be used.}

\item{toList}{An indicator of whether convert the UMAP representation to a list.}
}
\value{
Returns a UMAP representation.

Returns a revised iSCMEBObj object.

Returns a revised iSCMEBResObj object.

Returns the UMAP matrix.
}
\description{
The function \code{CalculateUMAP} is used to run UMAP dimensionality reduction.
}
\examples{
data(iSCMEBObj_toy)
library(Seurat)
iSCMEBObj_toy <- CalculateUMAP(iSCMEBObj_toy, reduction="iSCMEB")
## resList <- CalculateUMAP(iSCMEBObj_toy@resList, reduction="iSCMEB")
}
\seealso{
\code{\link{iSCMEBObj-class}}, \code{\link{fit.iscmeb}}, \code{\link{iSCMEBResObj-class}}
}
