% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Getexpression.r
\name{doDEG}
\alias{doDEG}
\title{Run DEG for a Seurat object.}
\usage{
doDEG(seu, topn = 10)
}
\arguments{
\item{seu}{A Seurat object.}

\item{topn}{The number of features to be extracted when \code{features} is not provided. 10 by default.}
}
\description{
The function \code{doDEG} is used to run DEG for a Seurat object
}
\seealso{
\code{\link{IntegrateSpaData}}
}
