% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Getexpression.r
\name{doHeatmap}
\alias{doHeatmap}
\title{Heatmap for spots-by-feature matrix}
\usage{
doHeatmap(
  seu,
  features = NULL,
  topn = 10,
  cell_label = "Domain",
  grp_label = FALSE,
  pt_size = 4,
  grp_color = NULL,
  ncol.legend = 1,
  ...
)
}
\arguments{
\item{seu}{A Seurat object.}

\item{features}{The features to be plotted.}

\item{topn}{The number of features to be extracted when \code{features} is not provided. 10 by default.}

\item{cell_label}{A string specify the name of legend. \code{"Domain"} by default.}

\item{grp_label}{An indicator of whether display the group names. \code{FALSE} by default.}

\item{pt_size}{The point size used in the plot. 4 by default.}

\item{grp_color}{The colors to use for the group color bar. \code{NULL} by default.}

\item{ncol.legend}{The number of columns of legend.}

\item{...}{Arguments passed to other methods.}
}
\description{
The function \code{doHeatmap} is used to plot heatmap for a Seurat object with expressioin data.
}
\seealso{
\code{\link{IntegrateSpaData}}
}
