% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SelectModel.r
\name{embeddings}
\alias{embeddings}
\alias{embeddings.iSCMEBObj}
\alias{embeddings.iSCMEBResObj}
\title{Get the identity of iSC.MEB}
\usage{
embeddings(obj)

\method{embeddings}{iSCMEBObj}(obj)

\method{embeddings}{iSCMEBResObj}(obj)
}
\arguments{
\item{obj}{An iSCMEBObj object or iSCMEBResObj object.}
}
\value{
Returns a list of embedding, whose i-th element is the embedding of i-th tissue section.

Returns a list of embedding, whose i-th element is the embedding of i-th tissue section.
}
\description{
The function \code{embeddings} is used to get the low dimensional embeddings of iSC.MEB model.
}
\details{
iSCMEBResObj is an object that contains all iSC.MEB solution information. It is the output of function \code{fit.iscmeb}, which is the main function of our package.

iSCMEBResObj is an object that contains all iSC.MEB solution information. It is the output of function \code{fit.iscmeb}, which is the body of our algorithm.
}
\examples{
data(iSCMEBObj_toy)
EmbeddingsList <- embeddings(iSCMEBObj_toy)
}
\seealso{
\code{\link{iSCMEBObj-class}}, \code{\link{fit.iscmeb}}
}
