% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{Human_HK_genes}
\alias{Human_HK_genes}
\title{Human housekeeping genes from the Housekeeping and Reference Transcript Atlas.}
\format{
A data.frame which includes the human housekeeping genes information in the columns named "Gene" and "Ensembl".
}
\usage{
data(Human_HK_genes)
}
\description{
Human housekeeping genes from the Housekeeping and Reference Transcript Atlas.
}
