% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SelectModel.r
\name{idents}
\alias{idents}
\alias{idents.iSCMEBObj}
\alias{idents.iSCMEBResObj}
\title{Get the identity of iSC.MEB}
\usage{
idents(obj)

\method{idents}{iSCMEBObj}(obj)

\method{idents}{iSCMEBResObj}(obj)
}
\arguments{
\item{obj}{An iSCMEBObj object or iSCMEBResObj object.}
}
\value{
Returns a list of identity, whose i-th element is the identity of i-th tissue section.

Returns a list of identity, whose i-th element is the identity of i-th tissue section.
}
\description{
The function \code{idents} is used to get the clustering result of iSC.MEB model.
}
\details{
iSCMEBResObj is an object that contains all iSC.MEB solution information. It is the output of function \code{fit.iscmeb}, which is the main function of our package.

iSCMEBResObj is an object that contains all iSC.MEB solution information. It is the output of function \code{fit.iscmeb}, which is the body of our algorithm.
}
\examples{
data(iSCMEBObj_toy)
IdentsList <- idents(iSCMEBObj_toy)
}
\seealso{
\code{\link{iSCMEBObj-class}}, \code{\link{fit.iscmeb}}
}
