% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Getexpression.r
\name{IntegrateSpaData}
\alias{IntegrateSpaData}
\title{Integrate multiple SRT expression data}
\usage{
IntegrateSpaData(
  obj,
  species = c("human", "mouse", "unknown"),
  custom_housekeep = NULL
)
}
\arguments{
\item{obj}{A iSCMEBObj object.}

\item{species}{An optional string, one of 'Human', 'Mouse' and 'Unknown', specify the species of the SRT data to help choose the housekeeping genes. 'Unknown' means only using the PRECAST results reconstruct the alligned gene expression.}

\item{custom_housekeep}{User-specified housekeeping genes.}
}
\description{
The function \code{IntegrateSpaData} is used to integrate multiple SRT data based on the iSCMEBObj by iSC.MEB model fitting.
}
\seealso{
\code{\link{iSCMEBObj-class}}
}
