% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{iSCMEB}
\alias{iSCMEB}
\title{Fit a iSC.MEB model.}
\usage{
iSCMEB(iSCMEBObj, K = NULL)
}
\arguments{
\item{iSCMEBObj}{A iSCMEBObj object.}

\item{K}{An optional integer or integer vector, specify the candidates of number of clusters. if \code{K=NULL}, it will be set to 5~12.}
}
\value{
Returns a revised iSCMEBObj object.
}
\description{
Fit a iSC.MEB model.
}
\details{
The model results are saved in the slot of resList.
}
\examples{
data(iSCMEBObj_toy)
library(Seurat)
iSCMEBObj_toy <- runPCA(iSCMEBObj_toy)
iSCMEBObj_toy <- CreateNeighbors(iSCMEBObj_toy, platform = "Visium")
iSCMEBObj_toy <- SetModelParameters(iSCMEBObj_toy, maxIter=10)
iSCMEBObj_toy <- iSCMEB(iSCMEBObj_toy, K = 7)
}
\seealso{
\code{\link{iSCMEBObj-class}}, \code{\link{runPCA}}, \code{\link{CreateNeighbors}}, \code{\link{SetModelParameters}}
}
