% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\docType{class}
\name{iSCMEBResObj-class}
\alias{iSCMEBResObj-class}
\title{Each iSCMEBResObj object has a number of slots which store information.}
\description{
Each iSCMEBResObj object has a number of slots which store information. Key slots to access are listed below.
\itemize{
  \item \code{posList} - The position matrix list for a iSCMEBResObj object.
  \item \code{paramList} - The model parameter settings for a iSCMEBResObj object.
  \item \code{fitList} - The details of iSC.MEB models.
  \item \code{project} - Name of the project.
  \item \code{reduction} - The dimension reduction result of iSC.MEB model.
  \item \code{idents} - The clustering result of iSC.MEB model.
}
}
