% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{Mouse_HK_genes}
\alias{Mouse_HK_genes}
\title{Mouse housekeeping genes from the Housekeeping and Reference Transcript Atlas.}
\format{
A data.frame which includes the mouse housekeeping genes information in the columns named "Gene" and "Ensembl".
}
\usage{
data(Mouse_HK_genes)
}
\description{
Mouse housekeeping genes from the Housekeeping and Reference Transcript Atlas.
}
