% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualization.r
\name{plot_RGB}
\alias{plot_RGB}
\title{Plot spatial RGB heatmap.}
\usage{
plot_RGB(position, embed_3d, pointsize = 2, ...)
}
\arguments{
\item{position}{A two columns matrix or data.fram, provides the coordinates information.}

\item{embed_3d}{A three columns  matrix or data.fram, provides the rgb information.}

\item{pointsize}{The size of point in the scatter plot.}

\item{...}{Arguments passed to other methods.}
}
\value{
Returns a RGB plot.
}
\description{
The function \code{plot_RGB} is used to plot spatial RGB heatmap.
}
\examples{
data(iSCMEBObj_toy)
library(Seurat)
position <- iSCMEBObj_toy@posList[[1]]
tsne3 <- iSCMEBObj_toy@resList@reduction$TSNE3[[1]]
p1 <- plot_RGB(position, tsne3)
}
