% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualization.r
\name{plot_scatter}
\alias{plot_scatter}
\title{Scatter plot for two-dimensional embeddings}
\usage{
plot_scatter(
  position,
  labels,
  label_name,
  axis_names = c("tSNE1", "tSNE2"),
  cols = NULL,
  no_guides = FALSE,
  do_density = FALSE,
  no_axis_name = FALSE,
  point_size = 2,
  point_alpha = 1,
  nrow.legend = NULL,
  legend.position = "bottom",
  no_axis = FALSE,
  ...
)
}
\arguments{
\item{position}{A two columns matrix or data.fram, provides the coordinates information.}

\item{labels}{A vector provides the label information.}

\item{label_name}{A string provides the name of labels.}

\item{axis_names}{Axis names.}

\item{cols}{A vector determine the colors used in the plot.}

\item{no_guides}{A indicator of whether remove the legend. FALSE by default.}

\item{do_density}{A indicator of whether plot the density. FALSE by default.}

\item{no_axis_name}{A indicator of whether remove axis names. FALSE by default.}

\item{point_size}{The point size of scatter plot. 2 by default.}

\item{point_alpha}{The transparency of scatter plot. 1 by default.}

\item{nrow.legend}{An integer specify row number in legend.}

\item{legend.position}{The position of legend. "bottom" by default.}

\item{no_axis}{A indicator of whether remove axis labels and names.}

\item{...}{Arguments passed to other methods.}
}
\value{
Returns a scatter plot.
}
\description{
The function \code{plot_scatter} is used to draw a scatter plot.
}
\examples{
data(iSCMEBObj_toy)
library(Seurat)
position <- iSCMEBObj_toy@posList[[1]]
labels <- idents(iSCMEBObj_toy)[[1]]
p1 <- plot_scatter(position, labels, "clusters")
}
