% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SelectModel.r
\name{SelectModel}
\alias{SelectModel}
\alias{SelectModel.ISCMEBObj}
\alias{SelectModel.SCMEB_Result_Object}
\title{Select best ISCMEB model from candidated models.}
\usage{
SelectModel(
  obj,
  criteria = c("MBIC", "MAIC", "BIC", "AIC"),
  c_penalty = 1,
  K = NULL
)

\method{SelectModel}{ISCMEBObj}(
  obj,
  criteria = c("MBIC", "MAIC", "BIC", "AIC"),
  c_penalty = 1,
  K = NULL
)

\method{SelectModel}{SCMEB_Result_Object}(
  obj,
  criteria = c("MBIC", "MAIC", "BIC", "AIC"),
  c_penalty = 1,
  K = NULL
)
}
\arguments{
\item{obj}{A ISCMEBObj object or SCMEB_Result_Object object.}

\item{criteria}{A string, specify the criteria used for selecting the number of clusters, supporting "MBIC", "MAIC", "BIC" and "AIC" ("MBIC" by default).}

\item{c_penalty}{An optional positive value, the adjusted constant used in the MBIC criteria (1 by default).}

\item{K}{An optional number of cluster. When K is not null, the ISCMEB solution with K clusters will be selected, default is \code{NULL}.}
}
\value{
Returns a revised ISCMEBObj object.

Returns a revised list
}
\description{
Select best ISCMEB model from candidated models.
}
\details{
SCMEB_Result_Object is an object that contains all ISCMEB solution information. It is the output of function SC.MEB2, which is the main function of our package.
}
\examples{
data(ISCMEBObj_toy)
library(Seurat)
## For convenience, we show the ISCMEBObj for perform dimension reduction. 
## Users can use PCA method or WPCA.
ISCMEBObj_toy2 <- runPCA(ISCMEBObj_toy)
## seulist <- ISCMEBObj_toy$seulist
## seulist <- runPCA(seulist)
}
\seealso{
\code{\link{ISCMEBObj-class}}, \code{\link{SC.MEB2}}
}
