% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualization.r
\name{SpaHeatMap}
\alias{SpaHeatMap}
\alias{SpaHeatMap.iSCMEBObj}
\alias{SpaHeatMap.iSCMEBResObj}
\alias{SpaHeatMap.list}
\title{Plot spatial heatmap}
\usage{
SpaHeatMap(
  obj,
  item = c("cluster", "batch"),
  plot_type = c("Scatter", "RGB_TSNE", "RGB_UMAP"),
  title_name = NULL,
  combine = TRUE,
  layout.dim = c(1, 2),
  common.legend = TRUE,
  ...
)

\method{SpaHeatMap}{iSCMEBObj}(
  obj,
  item = c("cluster", "batch"),
  plot_type = c("Scatter", "RGB_TSNE", "RGB_UMAP"),
  title_name = NULL,
  combine = TRUE,
  layout.dim = c(1, 2),
  common.legend = TRUE,
  ...
)

\method{SpaHeatMap}{iSCMEBResObj}(
  obj,
  item = c("cluster", "batch"),
  plot_type = c("Scatter", "RGB_TSNE", "RGB_UMAP"),
  title_name = NULL,
  combine = TRUE,
  layout.dim = c(1, 2),
  common.legend = TRUE,
  ...
)

\method{SpaHeatMap}{list}(
  obj,
  item = c("cluster", "batch"),
  plot_type = c("Scatter", "RGB_TSNE", "RGB_UMAP"),
  title_name = NULL,
  combine = TRUE,
  layout.dim = c(1, 2),
  common.legend = TRUE,
  ...
)
}
\arguments{
\item{obj}{A iSCMEBObj object or a iSCMEBResObj object.}

\item{item}{Which feature to be used in the plot.}

\item{plot_type}{A string shows the type of plot.}

\item{title_name}{A string shows the title of plot.}

\item{combine}{An indicator of whether plot all on a figure. If TRUE, all figures are plotted; otherwise, return a list with each plot as component. TRUE by default.}

\item{layout.dim}{The dimension in the layout of plots when \code{combine = TRUE}.}

\item{common.legend}{An indicator of whether combine the legend of multiple plots. TRUE by default.}

\item{...}{Arguments passed to other methods.}
}
\value{
Return a ggplot2 object or list of ggplots objects.

Return a ggplot2 object or list of ggplots objects.

Return a ggplot2 object or list of ggplots objects.
}
\description{
The function \code{SpaHeatMap} is used to visualize the physical location vs cluster or batch.
}
\examples{
data(iSCMEBObj_toy)
library(Seurat)
p1 <- SpaHeatMap(iSCMEBObj_toy, item="cluster", plot_type="Scatter", common.legend=TRUE)
p2 <- SpaHeatMap(iSCMEBObj_toy, item="cluster", plot_type="RGB_TSNE", common.legend=TRUE)
}
\seealso{
\code{\link{iSCMEBObj-class}}, \code{\link{plot_scatter}}, \code{\link{plot_RGB}}, \code{\link{fit.iscmeb}}, \code{\link{iSCMEBResObj-class}}
}
