% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SelectModel.r
\name{embeddings}
\alias{embeddings}
\alias{embeddings.iSCMEBObj}
\alias{embeddings.iSCSCMEB_Result_ObjectMEBObj}
\title{Get the identity of iSC.MEB}
\usage{
embeddings(obj)

\method{embeddings}{iSCMEBObj}(obj)

\method{embeddings}{iSCSCMEB_Result_ObjectMEBObj}(obj)
}
\arguments{
\item{obj}{An iSCMEBObj object or SCMEB_Result_Object object.}
}
\value{
Returns a list of embedding, whose i-th element is the embedding of i-th tissue section.

Returns a list of embedding, whose i-th element is the embedding of i-th tissue section.
}
\description{
Get the identity of iSC.MEB
}
\details{
SCMEB_Result_Object is an object that contains all ISCMEB solution information. It is the output of function \code{fit.iscmeb}, which is the main function of our package.
}
\examples{
data(iSCMEBObj_toy)
EmbeddingsList <- embeddings(iSCMEBObj_toy)
}
\seealso{
\code{\link{iSCMEBObj-class}}, \code{\link{fit.iscmeb}}
}
