% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{ISCMEB}
\alias{ISCMEB}
\title{Fit a ISCMEB model.}
\usage{
ISCMEB(ISCMEBObj, K = NULL)
}
\arguments{
\item{ISCMEBObj}{A ISCMEBObj object.}

\item{K}{An optional integer or integer vector, specify the candidates of number of clusters. if \code{K=NULL}, it will be set to 4~12.}
}
\value{
Returns a revised ISCMEBObj object.
}
\description{
Fit a ISCMEB model.
}
\details{
The model results are saved in the slot of resList.
}
\examples{
data(ISCMEBObj_toy)
library(Seurat)
ISCMEBObj_toy <- runPCA(ISCMEBObj_toy)
ISCMEBObj_toy <- CreateNeighbors(ISCMEBObj_toy, platform = "Visium")
ISCMEBObj_toy <- SetModelParameters(ISCMEBObj_toy, maxIter=10)
ISCMEBObj_toy <- ISCMEB(ISCMEBObj_toy, K = 7)
}
\seealso{
\code{\link{ISCMEBObj-class}}, \code{\link{runPCA}}, \code{\link{CreateNeighbors}}, \code{\link{SetModelParameters}}
}
