% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{iSCMEBObj_toy}
\alias{iSCMEBObj_toy}
\title{A toy data to help user get started with our packages.}
\format{
A iSCMEBObj. The \code{seulist} slot includes a list of two Seurat data.
}
\usage{
data(iSCMEBObj_toy)
}
\description{
A toy data to help user get started with our packages.
}
