% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunPCA.r
\name{runPCA}
\alias{runPCA}
\alias{runPCA.iSCMEBObj}
\alias{runPCA.list}
\title{Run a PCA dimensionality reduction.}
\usage{
runPCA(
  obj,
  npcs = 15,
  pca.method = c("APCA", "PCA", "WPCA"),
  reduction.name = "pca"
)

\method{runPCA}{iSCMEBObj}(
  obj,
  npcs = 15,
  pca.method = c("APCA", "PCA", "WPCA"),
  reduction.name = "pca"
)

\method{runPCA}{list}(
  obj,
  npcs = 15,
  pca.method = c("APCA", "PCA", "WPCA"),
  reduction.name = "pca"
)
}
\arguments{
\item{obj}{A seuList or iSCMEBObj object or matrix list.}

\item{npcs}{Total Number of PCs to compute and store (15 by default).}

\item{pca.method}{A string, specify which kind of PCA to be used. Supporting "APCA" (the approximate PCA), "PCA" (the classical PCA) and "WPCA" (the weighted PCA). Default method is APCA.}

\item{reduction.name}{Dimensional reduction name, pca by default}
}
\value{
Returns a revised iSCMEBObj object when input a iSCMEBObj.

Returns a revised list when input a seuList or matrix list.
}
\description{
Run a PCA dimensionality reduction.
}
\details{
seuList is a \link{list} with Seurat object as component, and each Seurat object includes the raw expression count matrix, spatial coordinates and meta data for each data batch, where the spatial coordinates information must be saved in the metadata of Seurat, named "row" and "col" for each data batch. matrix list is a list of log-transformed expression matrix, element of which has same columns.
}
\examples{
data(iSCMEBObj_toy)
library(Seurat)
## For convenience, we show the iSCMEBObj for perform dimension reduction. 
## Users can use PCA method or WPCA.
iSCMEBObj_toy2 <- runPCA(iSCMEBObj_toy)
## seulist <- iSCMEBObj_toy$seulist
## seulist <- runPCA(seulist)
}
\seealso{
\code{\link{iSCMEBObj-class}}
}
