% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SelectModel.r
\name{SelectModel}
\alias{SelectModel}
\alias{SelectModel.iSCMEBObj}
\alias{SelectModel.iSCMEBResObj}
\title{Select best iSC.MEB model from candidated models.}
\usage{
SelectModel(
  obj,
  criteria = c("MBIC", "MAIC", "BIC", "AIC"),
  c_penalty = 1,
  K = NULL
)

\method{SelectModel}{iSCMEBObj}(
  obj,
  criteria = c("MBIC", "MAIC", "BIC", "AIC"),
  c_penalty = 1,
  K = NULL
)

\method{SelectModel}{iSCMEBResObj}(
  obj,
  criteria = c("MBIC", "MAIC", "BIC", "AIC"),
  c_penalty = 1,
  K = NULL
)
}
\arguments{
\item{obj}{A iSCMEBObj object or iSCMEBResObj object.}

\item{criteria}{A string, specify the criteria used for selecting the number of clusters, supporting "MBIC", "MAIC", "BIC" and "AIC" ("MBIC" by default).}

\item{c_penalty}{An optional positive value, the adjusted constant used in the MBIC criteria (1 by default).}

\item{K}{An optional number of cluster. When K is not null, the iSC.MEB solution with K clusters will be selected, default is \code{NULL}.}
}
\value{
Returns a revised iSCMEBObj object.

Returns a revised list
}
\description{
The function \code{SelectModel} is used to select an optimal solution by given criteria or number of clusters.
}
\details{
iSCMEBResObj is an object that contains all iSC.MEB solution information. It is the output of function \code{fit.iscmeb}, which is the body of our algorithm.
}
\examples{
data(iSCMEBObj_toy)
iSCMEBObj_toy <- SelectModel(iSCMEBObj_toy)
}
\seealso{
\code{\link{iSCMEBObj-class}}, \code{\link{fit.iscmeb}}, \code{\link{iSCMEBResObj-class}}
}
